Moving utils outside the 'apis' folder.
